feat(json): Update jackson version to 2.6.3
related to CAM-5099

+++ spin/dataformat-json-jackson/src/test/java/org/camunda/spin/json/tree/JsonTreeMapObjectToJsonScriptTest.java
@@ -49,7 +49,7 @@ public abstract class JsonTreeMapObjectToJsonScriptTest extends ScriptTest {
     json = script.getVariable("integerVar");
 
     assertThat(json.isNumber()).isTrue();
-    assertThat(json.numberValue()).isEqualTo(42);
+    assertThat(json.numberValue().intValue()).isEqualTo(42);
 
     SpinJsonNode jsonList = script.getVariable("listVar");
